Moved Most Docker and HTTP Functions
